pyproject.toml
[tool.poetry]
name = "course-project"
version = "0.1.0"
description = ""
authors = ["EgorovNikolay <egorovnikolay14021994@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pandas = "^2.2.3"
openpyxl = "^3.1.5"
requests = "^2.32.3"
python-dotenv = "^1.0.1"


[tool.poetry.group.lint.dependencies]
flake8 = "^7.1.1"
isort = "^6.0.0"
black = "^25.1.0"
mypy = "^1.14.1"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"
pytest-cov = "^6.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 119
exclude = '''
/(
    \.git
)
'''

[tool.isort]
line_length = 119

[tool.mypy]
disallow_untyped_defs = true
warn_return_any = true
exclude = 'venv'
